Add user matcher code to get_or_create_lti_user
The general idea is to subclass `LTIUserMatcher` and provide custom
logic to search for an existing user based on LTI launch data.

The custom matcher class names can be specified as the
matcher_class_name for a given LTIToolConsumer instance.  If no
matcher is specified, users are only looked up based on the
combination of consumer key and LTI user_id.